// src/client/utils/interfaces.ts
export interface IChirp {
    id: number;
    userid: number;
    chirp: string;
    created_at: Date;
    name: string;
}

export interface IUser {
    id: number;
    name: string;
}














//Half Stack Class based
// export interface IChirp{
//     id: number;
//     name: string;
//     text: string;
//     written_at: Date;

// }